refactor: major architectural changes. - uses parking_lot for mutex and rwlock - changed key type from u32 to u64 (preparation to design it like sqlite's row_id internal type) - added a method to find page_id for the given target key, or return the rightmost_child_id on curr internal   node

+++ src/error.rs
@@ -21,8 +21,8 @@ pub enum DbError {
     LruEviction,
 
     #[error("key already exists, cannot insert again: key={0}")]
-    DuplicateKey(u32),
+    DuplicateKey(u64),
 
     #[error("key does not exists: key={0}")]
-    KeyNotFound(u32),
+    KeyNotFound(u64),
 }
